src: add experimental secure heap
This adds an experimental implementation of what OpenSSL calls a "secure
heap". At its core, this is a dynamic memory implementation with certain
protection features. If possible, protected memory should never be
written to the disk ("locked in memory"), it should not be contained in
core dumps and it should automatically erase its contents when no longer
needed.